Add Preprocessed implementation for SemanticDomain
Though it doesn't offer a performance optimization, since reconciler didn't execute the same getValid* method twice with the same argument
